Adding the String Resource Endpoint Let's finish this endpoint. At the bottom of ProgrammerController, add a new public function editTaglineAction(). We already know that the route should be /api/programmers/{nickname}/tagline. To be super hip, add an @Method annotation: we know this should only match PUT requests.
Like before, type-hint the Programmer argument so that Doctrine will query for it for us, using the nickname value. And, we'll also need the Request argument. I could use a form like before... but this is just so simple: {dollar_sign}programmer->setTagLine({dollar_sign}request->getContent()).
Literally: read the text from the request body and set that on the programmer.

Now, save: {dollar_sign}em = {dollar_sign}this->getDoctrine()->getManager(), {dollar_sign}em->persist({dollar_sign}programmer) and {dollar_sign}em->flush(). For the return, it's not JSON! Return a plain new Response() with {dollar_sign}programmer->getTagLine(), a 200 status code, and a Content-Type header of text/plain.

Now, this is a good-looking, super-weird endpoint. Copy the test method name and try it out:

./vendor/bin/phpunit --filter testEditTagLine src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
We're green! Next, let's look at a weirder endpoint.